ADDON-34283: ucc_ui_lib: updated configs and structure to have single entry point js and template file
