+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/resources/splunk/appserver/templates/base.html
@@ -8,7 +8,7 @@ app_name = cherrypy.request.path_info.split('/')[3]
 <head>
     <meta charset="utf-8">
     <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title>${_('Inputs')}</title>
+    <title>${_('Loading')}</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="apple-touch-icon" href="apple-touch-icon.png">
 </head>
@@ -22,7 +22,7 @@ app_name = cherrypy.request.path_info.split('/')[3]
     </script>
 
     <%
-    page_path = "/static/app/" + app_name + "/js/build/" + page + ".js"
+    page_path = "/static/app/" + app_name + "/js/build/entry_page.js"
     %>
 
     <script src="${make_url(page_path)}"></script>
