+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/resources/splunk/default/data/ui/views/configuration.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0"?>
-<view template="${package.name}:/templates/configuration.html" type="html" isDashboard="False">
+<view template="${package.name}:/templates/base.html" type="html" isDashboard="False">
 	<label>Configuration</label>
 </view>
