+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/resources/splunk/default/data/ui/views/inputs.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0"?>
-<view template="${package.name}:/templates/inputs.html" type="html" isDashboard="False">
+<view template="${package.name}:/templates/base.html" type="html" isDashboard="False">
 	<label>Inputs</label>
 </view>
