--- splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/pages/inputs/StartStyles.js
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-import { variables, mixins } from '@splunk/themes';
-
-const StyledContainer = styled.div`
-    ${mixins.reset('inline')};
-    display: block;
-    font-size: ${variables.fontSizeLarge};
-    line-height: 200%;
-    margin: calc(${variables.spacing} * 2) calc(${variables.spacing} * 2);
-`;
-
-const StyledGreeting = styled.div`
-    font-weight: bold;
-    color: ${variables.infoColor};
-    font-size: ${variables.fontSizeXXLarge};
-`;
-
-export { StyledContainer, StyledGreeting };
