+++ splunk_add_on_ucc_framework/ucc_ui_lib/webpack.config.js
@@ -4,20 +4,11 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 const webpackMerge = require('webpack-merge');
 const baseConfig = require('@splunk/webpack-configs/base.config').default;
 
-// Set up an entry config by iterating over the files in the pages directory.
-const entries = fs
-    .readdirSync(path.join(__dirname, 'src/main/webapp/pages'))
-    .filter((pageFile) => !/^\./.test(pageFile))
-    .reduce((accum, page) => {
-        accum[page] = path.join(__dirname, 'src/main/webapp/pages', page);
-        return accum;
-    }, {});
-
 module.exports = webpackMerge(baseConfig, {
-    entry: entries,
+    entry: path.join(__dirname, 'src/main/webapp/pages/entry_page'),
     output: {
         path: path.join(__dirname, 'stage/appserver/static/js/build'),
-        filename: '[name].js',
+        filename: 'entry_page.js',
     },
     plugins: [
         new CopyWebpackPlugin([
